arc/requirements.txt
Adafruit-Blinka==5.4.0
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==2.17.0
Adafruit-PureIO==1.1.5
aiohttp==3.7.3
amqp==2.6.1
APScheduler==3.7.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorzero==1.1
Django==3.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-celery-beat==2.0.0
django-chartjs==2.2.1
django-crispy-forms==1.9.2
django-forms-bootstrap==3.1.0
django-timedeltafield==0.7.10
django-timezone-field==4.0
gpiozero==1.5.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
imutils==0.5.3
kombu==4.6.11
multidict==5.1.0
numpy==1.19.2
opencv-contrib-python==4.1.0.25
pkg-resources==0.0.0
prompt-toolkit==3.0.13
psycopg2==2.8.6
pyftdi==0.51.2
pymongo==3.11.3
pyserial==3.4
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pyusb==1.1.0
requests==2.25.1
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.23
sqlparse==0.3.1
sysv-ipc==1.0.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.1.0
